fix: add wallpapers to search results
